wyprzedzanie, wykorzystanie kata w PID, fixy

+++ stm32-atollic/SelfieF7/Src/VL53L0X/Czujniki.h
@@ -17,7 +17,7 @@ extern xSemaphoreHandle i2c2_semaphore;
 uint16_t range[VLX_SENSOR_COUNT];
 uint16_t lane_change_treshold;
 
-uint8_t flags[3];
+uint16_t flags[3];
 
 void StartCzujnikiTask(void const * argument);
